requirements.txt
selenium
instapy
python-dotenv
textblob==0.15.3
